+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes after its initial development up to June 2020 (v1.0) are documented in this file.
 
+### v1.3 (May 2021)
+    - Correct bug in exe_wrapper if pid=True.
+    - Allow non-numeric parameters in std_io functions.
+
 ### v1.2 (Dec 2020)
     - Removed scipy.optimize from tests of wrappers.
     - Use build instead of cibuildwheel to build pure Python wheels.